// Test/src/app/models/masternumber-data.ts
export class MasterNumberData{
   mastNumId: number;
   actualMastNumId: string;
   catCode: string;
   createdBy: number;
   createdDatetime: string;
   customerId: number;
   customerName: String;
   company: string;
   address: string;
   phoneNo:string;
   lossAddress: string;
   lossAddressCity: string;
   lossAddressState: string;
   lossAddressZip: string;
   billingAddress: string;
   billingAddressCity: string;
   billingAddressState: string;
   billingAddressZip: string;
   billingAddressSame: boolean;
   city: string;
   zip: string;
   tenantName: string;
   tenantNumber: string;
   damageType: string;
   duplicateComments: string;
   duplicateEnteredBy: number;
   duplicateEnteredDatetime: string;
   isDuplicate: string;
   isInsured: string;
   mastNumCode: string;
   mastNumDesc: string;
   mastNumHash: string;
   propertyType: string;
   restCompLocationId: number;
   restorationType: string;
   masterName: string;
   companyName: string;
   responsibleParty: string;
   sourceId: number;
   sourceType: number;
   programType: string;
   causeOfLoss: string;
   yearBuilt: string;
   status: string;
   updatedBy: number;
   updatedDatetime: string;
   
}